+++ CHANGELOG.md
@@ -1,1 +1,10 @@
-- ♻️  A bit of cleaning on initUserAssociation
+- ✨ Now we need is_site modification on association.
+- 💄 Added button to charters in dashboard association.
+- 💄 Added charter date in association dashboard.
+- 🗃️ Changed logos layout.
+- 💄 Added amount earned in project dashboard.
+- ✨ Added budget-file to Project list.
+- 💄 Added download-recap button on project detail view.
+- 🗃️ Fields on project cannot equal 0.
+- 💄 "Return to top" button is at the bottom.
+- 💬 Updated subtitle.